requirements.txt
streamlit==1.51.0
typesense
pandas
google-genai
langchain-google-genai
python-dotenv
pydantic-settings
fastapi
uvicorn
slowapi
requests
watchdog
jinja2
python-multipart
markdown
bcrypt
Pillow
pytest